move loads functionality from resolver class to has_arguments module
